refactor(user-service): relocate service interfaces to dedicated impl package for better structure
